Streamlit Web App/requirements.txt
numpy==2.1.3
Pillow==11.0.0
streamlit==1.37.1
tensorflow==2.17.0
